services/python/requirements.txt
fastapi
uvicorn[standard]
python-multipart
pymupdf
sentence-transformers
llama-index-core
numpy
python-dotenv
passlib[bcrypt]
python-jose[cryptography]
lancedb
pytest
pytest-asyncio
pytest-mock
